+++ src/error.rs
@@ -4,24 +4,27 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-/// Defines all possible error variants for SAFE
+/// Defines all possible error variants for the SAFE library.
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Error {
-    /// A call to `squeeze`, `absorb` or `finish` that doesn't fit the
-    /// io-pattern.
+    /// This error occurs when the expected IO-pattern sequence wasn't followed
+    /// during the usage of the sponge algorithm.
     IOPatternViolation,
 
-    /// Invalid io-pattern.
+    /// This error occurs when the provided IO-pattern is not valid.
+    /// This means that:
+    /// - It doesn't start with a call to squeeze or
+    /// - It doesn't end with a call to absorb or
+    /// - Every call to absorb or squeeze has a positive length.
     InvalidIOPattern,
 
-    /// The input doesn't yield enough input elements.
+    /// This error occurs when the input elements provided to the
+    /// [`Sponge::absorb`] are less than the amount that should be absorbed.
     TooFewInputElements,
 
-    /// Failed to encrypt the message into the cipher with the provided secret
-    /// and nonce.
+    /// This error indicates a failure during the encryption process.
     EncryptionFailed,
 
-    /// Failed to decrypt the message from the cipher with the provided secret
-    /// and nonce.
+    /// This error indicates a failure during the decryption process.
     DecryptionFailed,
 }