+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add authenticated encryption and decryption [#6]
+- Add check for `cipher.len == message.len + 1` in `encrypt` and `decrypt` [#9]
 
 ### Changed
 
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add documentation
 
 <!-- ISSUES -->
+[#9]: https://github.com/dusk-network/safe/issues/9
 [#6]: https://github.com/dusk-network/safe/issues/6
 [#4]: https://github.com/dusk-network/safe/issues/4
 [#3]: https://github.com/dusk-network/safe/issues/3