+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Let `Sponge::start` take the io-pattern as `impl Into<Vec<Call>>` [#4]
 - Change `nonce` to be `&T` instead of `T` in `encrypt` and `decrypt` [#9]
+- Improve crate documentation [#13]
 - Rename `Encryption::assert_equal` to `Encryption::is_equal` [#15]
 
 ## [0.1.0] - 2024-03-07
@@ -32,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ISSUES -->
 [#15]: https://github.com/dusk-network/safe/issues/15
+[#13]: https://github.com/dusk-network/safe/issues/13
 [#9]: https://github.com/dusk-network/safe/issues/9
 [#6]: https://github.com/dusk-network/safe/issues/6
 [#4]: https://github.com/dusk-network/safe/issues/4